+++ test/e2e/_test.setup.ts
@@ -22,9 +22,9 @@ import { randomUUID } from 'crypto';
 import archiver from 'archiver';
 
 // Local imports
-import { createLogger } from '../../test-runner/logger';
-import { Application, Logger, PositronPythonFixtures, PositronRFixtures, PositronUserSettingsFixtures, UserSetting } from '../../automation';
-import { createApp } from '../../smoke/src/utils';
+import { createLogger } from '../test-runner/logger';
+import { Application, Logger, PositronPythonFixtures, PositronRFixtures, PositronUserSettingsFixtures, UserSetting } from '../automation';
+import { createApp } from '../smoke/src/utils';
 
 const TEMP_DIR = `temp-${randomUUID()}`;
 const ROOT_PATH = process.cwd();
