+++ test/e2e/console/console-ansi.test.ts
