+++ test/e2e/console/console-autocomplete.test.ts
