+++ test/e2e/console/console-clipboard.test.ts
@@ -5,7 +5,7 @@
 
 import * as os from 'os';
 import { test, expect } from '../_test.setup';
-import { Application } from '../../../automation';
+import { Application } from '../../automation';
 
 test.use({
 	suiteId: __filename
