+++ test/e2e/console/console-history.test.ts
