+++ test/e2e/console/console-input.test.ts
