+++ test/e2e/console/console-output.test.ts
