+++ test/e2e/data-explorer/duckdb-sparklines.test.ts
