+++ test/e2e/data-explorer/helpers/100x100.ts
@@ -5,7 +5,7 @@
 
 import { join } from 'path';
 import * as fs from 'fs';
-import { Application } from '../../../../automation';
+import { Application } from '../../../automation';
 import { expect } from '@playwright/test';
 
 export const testDataExplorer = async (
