+++ test/e2e/data-explorer/large-data-frame.test.ts
