+++ test/e2e/output/console-ouput-log.test.ts
