+++ test/e2e/plots/autos.png
