+++ test/e2e/plots/graphviz.png
