+++ test/e2e/plots/plots.test.ts
@@ -9,7 +9,7 @@ import { ComparisonOptions } from 'resemblejs';
 import * as fs from 'fs';
 import { fail } from 'assert';
 import { test, expect } from '../_test.setup';
-import { Application } from '../../../automation';
+import { Application } from '../../automation';
 
 
 test.use({
