+++ test/e2e/plots/pythonScatterplot.png
