+++ test/e2e/variables/variables-expanded.test.ts
