+++ test/e2e/variables/variables-notebook.test.ts
