+++ test/e2e/variables/variables-pane.test.ts
